# backend/static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Movie Ratings Explorer API",
    "version": "2.0.0",
    "description": "RESTful API for the Optamo Movie Ratings platform. All list endpoints return a paginated envelope with server-side sorting support."
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Movie": {
        "type": "object",
        "properties": {
          "name":              { "type": "string",  "example": "The Godfather" },
          "genre":             { "type": "array", "items": { "type": "string" }, "example": ["Crime", "Drama"] },
          "release_year":      { "type": "integer", "example": 1972, "nullable": true },
          "number_user_rated": { "type": "integer", "example": 52498 },
          "average_rating":    { "type": "number",  "example": 4.32 }
        }
      },
      "PaginatedMovieResponse": {
        "type": "object",
        "properties": {
          "data":     { "type": "array", "items": { "$ref": "#/components/schemas/Movie" } },
          "total":    { "type": "integer", "description": "Total matching rows across all pages", "example": 95 },
          "page":     { "type": "integer", "description": "Current 1-indexed page number",       "example": 1 },
          "per_page": { "type": "integer", "description": "Number of results per page",          "example": 25 },
          "pages":    { "type": "integer", "description": "Total number of available pages",     "example": 4 }
        }
      }
    },
    "parameters": {
      "page": {
        "name": "page", "in": "query", "required": false,
        "schema": { "type": "integer", "default": 1, "minimum": 1 },
        "description": "1-indexed page number."
      },
      "per_page": {
        "name": "per_page", "in": "query", "required": false,
        "schema": { "type": "integer", "default": 25, "minimum": 1, "maximum": 100 },
        "description": "Number of results per page. Clamped server-side to 100."
      },
      "sort_by": {
        "name": "sort_by", "in": "query", "required": false,
        "schema": { "type": "string", "enum": ["name", "release_year", "average_rating", "number_user_rated"] },
        "description": "Column to sort results by."
      },
      "sort_dir": {
        "name": "sort_dir", "in": "query", "required": false,
        "schema": { "type": "string", "enum": ["asc", "desc"] },
        "description": "Sort direction."
      }
    }
  },
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": { "description": "Server is running and database file exists." }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login — returns a JWT (demo: admin / password)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string", "example": "admin" },
                  "password": { "type": "string", "example": "password" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Returns { token: string }" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/api/movies/top-rated": {
      "get": {
        "summary": "Paginated top-rated movies (>10K ratings AND avg > 4.0)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/page" },
          { "$ref": "#/components/parameters/per_page" },
          {
            "$ref": "#/components/parameters/sort_by",
            "schema": { "type": "string", "enum": ["name", "release_year", "average_rating", "number_user_rated"], "default": "average_rating" }
          },
          {
            "$ref": "#/components/parameters/sort_dir",
            "schema": { "type": "string", "enum": ["asc", "desc"], "default": "desc" }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated envelope of top-rated movies.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaginatedMovieResponse" }
              }
            }
          },
          "401": { "description": "Missing or invalid JWT token" }
        }
      }
    },
    "/api/movies/details": {
      "get": {
        "summary": "Movie details by name (single result)",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "name", "in": "query", "required": true, "schema": { "type": "string", "example": "The Godfather" } }
        ],
        "responses": {
          "200": {
            "description": "Single Movie object.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Movie" }
              }
            }
          },
          "400": { "description": "name query parameter is missing" },
          "404": { "description": "No movie found matching the name" },
          "401": { "description": "Missing or invalid JWT token" }
        }
      }
    },
    "/api/movies/search": {
      "get": {
        "summary": "Paginated movie search with multi-parameter filters",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "name",   "in": "query", "required": false, "schema": { "type": "string" }, "description": "Partial movie name (case-insensitive)." },
          { "name": "genre",  "in": "query", "required": false, "schema": { "type": "string", "example": "Drama" }, "description": "Genre filter." },
          { "name": "year",   "in": "query", "required": false, "schema": { "type": "integer", "example": 1994 }, "description": "Exact release year." },
          { "name": "rating", "in": "query", "required": false, "schema": { "type": "number",  "example": 4.0 }, "description": "Minimum average rating." },
          { "$ref": "#/components/parameters/page" },
          { "$ref": "#/components/parameters/per_page" },
          {
            "$ref": "#/components/parameters/sort_by",
            "schema": { "type": "string", "enum": ["name", "release_year", "average_rating", "number_user_rated"], "default": "name" }
          },
          {
            "$ref": "#/components/parameters/sort_dir",
            "schema": { "type": "string", "enum": ["asc", "desc"], "default": "asc" }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated envelope of matching movies.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaginatedMovieResponse" }
              }
            }
          },
          "401": { "description": "Missing or invalid JWT token" }
        }
      }
    }
  }
}